refactor(endpoints): modularize slack and discord endpoint helpers
